[test] -- refactoring, complete function set_eth
