+++ kernel/Makefile
@@ -1,6 +1,7 @@
 KDIR := /lib/modules/$(shell uname -r)/build
-#ccflags-y += -DIPE_DEBUG=1 -Wall 
-obj-m += nlKernRec.o
+ccflags-y += -DIPE_DEBUG=1 -Wall 
+obj-m += ipe.o 
+ipe-y = ipeDrv.o ipeDebug.o 
 
 all:
 	$(MAKE) -C $(KDIR) SUBDIRS=$(PWD) modules
